producer/requirements.txt
quixstreams==2.9.0
python-dotenv
zstandard
tqdm
pandas